Removed onsets/offsets from GuitarSet and moved note decoding function.
