+++ examples/tabcnn.py
@@ -5,7 +5,7 @@ from amt_tools.datasets import GuitarSet
 from amt_tools.models import TabCNN
 from amt_tools.features import CQT
 
-from amt_tools.train import train
+from amt_tools.train import train, validate
 from amt_tools.transcribe import *
 from amt_tools.evaluate import *
